refactor(feat/resource_information): move resource components to components/resources folder
