+++ pages/resources/[resourceId].js
@@ -1,8 +1,8 @@
 import React from "react";
 import { useRouter } from "next/router";
-import NewResourceComment from "../../components/resources/NewResourceComment";
-import ResourceInformation from '../../components/ResourceInformation';
 import { Box, CircularProgress } from "@mui/material";
+import NewResourceComment from "../../components/resources/NewResourceComment";
+import ResourceInformation from "../../components/resources/ResourceInformation";
 
 export default function Resources() {
   const { query, isReady } = useRouter();
